Add xml:base requirement to corpus.xml constraint
and remove the constraint (warning) from the idno spec.

The xml:base attribute is used  in combination with type="URI" in
queries in the dutil:get-corpus and dutil:get-corpus-info functions
of the DraCor API. So both are necessary for the corpus name to be
found. Therefore the constraint should be expressed in one place
and as critical.