Implement schematron rule checking either xml-model or /TEI/@type
This rule requires the a DraCor document to either reference the
schema in a xml-model processing instruction or adding a
type="dracor" to the root element. Having a type attribute allows
us to constrain the document type association in the DraCor Oxygen
framework to DraCor documents only and thus allow it to coexist
with the TEI P5 and other frameworks.