Revise schema for corpus.xml
In order to make corpus.xml TEI compliant without overriding the content
model of teiCorpus this defines a new root element dracorCorpus. It has
exactly one teiHeader child element.

In addition we simplifz the declaration of the corpus name with an
untyped idno element and use the more appropriate ref instead of idno
to declare the repository URL.